% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{board_size}
\alias{board_size}
\title{Board Size}
\usage{
board_size(b)
}
\arguments{
\item{b}{A square matrix representing a boggle board.}
}
\value{
The size of the board.
}
\description{
Return the size of a square boggle board.
}
\examples{
board <- random_board()
board_size(board)

}

