% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{index_of_neighbor}
\alias{index_of_neighbor}
\title{Index of neighbor}
\usage{
index_of_neighbor(idx, nr, bs)
}
\arguments{
\item{idx}{The index of the matrix as a vector
starting with the top left and proceeding down then right.}

\item{nr}{relative cell reference.  If a cell is at position
idx, the corresponding cell references are as follows.

\tabular{rrrrr}{
  8 \tab 7   \tab   6\cr
  5 \tab idx \tab   4\cr
  3 \tab 2   \tab   1\cr
}}

\item{bs}{Board size.}
}
\value{
Returns a the index of a neighboring cell.
}
\description{
This function could probably be minimized or done away with
by encoding the data differently when checking neighboring
cells.
}
\examples{
board        <- random_board()
bs           <- board_size(board)

# This example returns 5.  The index of the first cell in the
# second column of a 4x4 matrix.  The first arg (idx) is set
# to 1, representing the first cell in the first row.  The
# second argument (nr) which is set to 4 indicating the cell
# to the immediate right (east).  Finally the board size is set
# to 4.
index_of_neighbor(1, 4, 4)

}

