% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{is_word}
\alias{is_word}
\title{Is string a word?}
\usage{
is_word(letters)
}
\arguments{
\item{letters}{The character string being tested.}
}
\value{
A boolean indcating whether or not the character string
passed in as an argument is a word.
}
\description{
Returns a boolean value indicating whether the specified
character string is a word
}
\examples{
is_word('wow')
}

