% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{neighbors}
\alias{neighbors}
\title{Neighbors}
\usage{
neighbors(padded_board)
}
\arguments{
\item{padded_board}{A square matrix representing a boggle board surrounded by NA cells.}
}
\value{
an adjacency matrix where each column represents one
cell in the boggle board matrix, and each row represents one
of 8 possible adjacent cells.
}
\description{
Returns a matrix representing all neighbors for a given board
}
\examples{
board        <- random_board()
padded_board <- cbind(NA,rbind(NA,board,NA),NA)
neighbors(padded_board)

}

