% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{random_board}
\alias{random_board}
\title{Generate a new boggle board}
\usage{
random_board(size = 4, dice = boggle_dice)
}
\arguments{
\item{size}{Dimensions of the boggle board.  Default is 4 x 4.}

\item{dice}{A data frame containing the sides of each dice used in the game.  Default is the 16 six-sided dice data frame "boggle_dice"}
}
\value{
a matrix containing a new boggle board
}
\description{
Effectively "shake up" a new boggle board which results in a set
of dice being rolled and place in a square grid.
}
\examples{
random_board()
random_board(size = 5)

}

