% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{score_word}
\alias{score_word}
\title{Score word}
\usage{
score_word(word)
}
\arguments{
\item{word}{The word being scored.}
}
\value{
The score for the word passed in as an argument.
}
\description{
Returns the score for a given word.
}
\details{
Scoring follows standard boggle rules:

\itemize{
 \item{4 or fewer letters:  }{ 1 point}
 \item{5-letter words:      }{ 2 points}
 \item{6-letter words:      }{ 3 points}
 \item{7-letter words:      }{ 5 points}
 \item{7 letters or longer: }{ 11 points}
}

Also takes into consideration the "Qu" face on the dice which
represents a single character.
}
\examples{
score_word('CAT')
}

