% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{starts_with}
\alias{starts_with}
\title{Starts with}
\usage{
starts_with(letters)
}
\arguments{
\item{letters}{The character string being tested.}
}
\value{
A list of words that match the beginning of the character string.
}
\description{
Returns a list of words that have the specified character string
at the beginning of the word.
}
\examples{
starts_with('wow')
}

