% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boggle_solver.R
\name{word_search}
\alias{word_search}
\title{Word search}
\usage{
word_search(idx, fragment, indices, results, board, nb)
}
\arguments{
\item{idx}{The cell where the character string (fragment) originates.}

\item{fragment}{A character string being tested.}

\item{indices}{The indexes of the cells of the character string being tested.}

\item{results}{A list of words and indices.}

\item{board}{The scrabble board in play.}

\item{nb}{The matrix representing neighboring cells.}

\item{letters}{The character string being tested.}
}
\value{
A list of words and their indices on the boggle board.
}
\description{
This function recursively searches for all words available
when starting with a given cell.
}
\examples{
# Search for all words originating with the first upper
# left-hand cell.
board        <- random_board()
padded_board <- cbind(NA,rbind(NA,board,NA),NA)
nb           <- neighbors(padded_board)
results      <- list(words=c(),indices=c())
idx          <- 1
word_search(idx, board[idx], c(idx), results, board, nb)

}

